+++ src/main/groovy/com/ofg/infrastructure/config/WebAppConfiguration.groovy
@@ -1,7 +1,7 @@
 package com.ofg.infrastructure.config
 import com.ofg.infrastructure.discovery.ServiceDiscoveryConfiguration
 import com.ofg.infrastructure.healthcheck.HealthCheckConfiguration
-import com.ofg.infrastructure.metrics.registry.MetricsRegistryConfiguration
+import com.ofg.infrastructure.metrics.registry.MetricsConfiguration
 import com.ofg.infrastructure.web.swagger.SwaggerConfiguration
 import com.ofg.infrastructure.web.config.WebInfrastructureConfiguration
 import groovy.transform.CompileStatic
@@ -15,19 +15,19 @@ import org.springframework.context.annotation.Import
  * <ul>
  *  <li>{@link WebInfrastructureConfiguration} - contains configurations related to filter, service communication and web application setup</li>
  *  <li>{@link ServiceDiscoveryConfiguration} - contains configurations related to service discovery</li>
- *  <li>{@link MetricsRegistryConfiguration} - contains configurations with registry of metrics instances
+ *  <li>{@link MetricsConfiguration} - contains configurations with registry of metrics instances
  *  <li>{@link HealthCheckConfiguration} - contains configurations related to Health check verification
  *  <li>{@link SwaggerConfiguration} - contains configurations related to Swagger API documentation
  * </ul>
  *
  * @see WebInfrastructureConfiguration
  * @see ServiceDiscoveryConfiguration
- * @see MetricsRegistryConfiguration
+ * @see MetricsConfiguration
  * @see HealthCheckConfiguration
  * @see SwaggerConfiguration
  */
 @Configuration
 @CompileStatic
-@Import([WebInfrastructureConfiguration, ServiceDiscoveryConfiguration, MetricsRegistryConfiguration, HealthCheckConfiguration, SwaggerConfiguration])
+@Import([WebInfrastructureConfiguration, ServiceDiscoveryConfiguration, MetricsConfiguration, HealthCheckConfiguration, SwaggerConfiguration])
 class WebAppConfiguration {
 }